requirement.txt
numpy
scipy
matplotlib
seaborn
jupyter
tqdm
pillow
transformers==4.34.1
einops
absl-py
ml_collections
requests
mlxu==0.1.11
pydantic
fastapi
uvicorn
gradio
opencv-python-headless
scikit-video
scikit-image
natsort
scikit-learn
tadpak